+++ lib/rspec_tracer/cache.rb
@@ -17,6 +17,7 @@ module RSpecTracer
       @pending_examples = Set.new
       @all_files = {}
       @dependency = Hash.new { |hash, key| hash[key] = Set.new }
+      @examples_coverage = {}
     end
 
     def load_cache_for_run
@@ -50,23 +51,25 @@ module RSpecTracer
     end
 
     def cached_examples_coverage
-      return @examples_coverage if defined?(@examples_coverage)
+      return @examples_coverage if @examples_coverage_loaded
+
+      @examples_coverage_loaded = true
 
       cache_path = RSpecTracer.cache_path
       cache_path = File.dirname(cache_path) if RSpecTracer.parallel_tests?
       run_id = last_run_id(cache_path)
 
-      return @examples_coverage = {} if run_id.nil?
+      return @examples_coverage if run_id.nil?
 
       starting = Process.clock_gettime(Process::CLOCK_MONOTONIC)
       cache_dir = File.join(cache_path, run_id)
-      coverage = load_examples_coverage_cache(cache_dir)
+      load_examples_coverage_cache(cache_dir)
       ending = Process.clock_gettime(Process::CLOCK_MONOTONIC)
       elapsed = RSpecTracer::TimeFormatter.format_time(ending - starting)
 
       puts "RSpec tracer loaded cached examples coverage (took #{elapsed})" if RSpecTracer.verbose?
 
-      coverage
+      @examples_coverage
     end
 
     private
@@ -76,7 +79,7 @@ module RSpecTracer
 
       return unless File.file?(file_name)
 
-      JSON.parse(File.read(file_name))['run_id']
+      JSON.parse(File.read(file_name, encoding: 'UTF-8'))['run_id']
     end
 
     def load_all_examples_cache(cache_dir, discard_run_reason: true)
@@ -84,7 +87,7 @@ module RSpecTracer
 
       return unless File.file?(file_name)
 
-      @all_examples = JSON.parse(File.read(file_name)).transform_values do |examples|
+      @all_examples = JSON.parse(File.read(file_name, encoding: 'UTF-8')).transform_values do |examples|
         examples.transform_keys(&:to_sym)
       end
 
@@ -99,7 +102,7 @@ module RSpecTracer
 
       return unless File.file?(file_name)
 
-      @duplicate_examples = JSON.parse(File.read(file_name)).transform_values do |examples|
+      @duplicate_examples = JSON.parse(File.read(file_name, encoding: 'UTF-8')).transform_values do |examples|
         examples.map { |example| example.transform_keys(&:to_sym) }
       end
     end
@@ -109,7 +112,7 @@ module RSpecTracer
 
       return unless File.file?(file_name)
 
-      @interrupted_examples = JSON.parse(File.read(file_name)).to_set
+      @interrupted_examples = JSON.parse(File.read(file_name, encoding: 'UTF-8')).to_set
     end
 
     def load_flaky_examples_cache(cache_dir)
@@ -117,7 +120,7 @@ module RSpecTracer
 
       return unless File.file?(file_name)
 
-      @flaky_examples = JSON.parse(File.read(file_name)).to_set
+      @flaky_examples = JSON.parse(File.read(file_name, encoding: 'UTF-8')).to_set
     end
 
     def load_failed_examples_cache(cache_dir)
@@ -125,7 +128,7 @@ module RSpecTracer
 
       return unless File.file?(file_name)
 
-      @failed_examples = JSON.parse(File.read(file_name)).to_set
+      @failed_examples = JSON.parse(File.read(file_name, encoding: 'UTF-8')).to_set
     end
 
     def load_pending_examples_cache(cache_dir)
@@ -133,7 +136,7 @@ module RSpecTracer
 
       return unless File.file?(file_name)
 
-      @pending_examples = JSON.parse(File.read(file_name)).to_set
+      @pending_examples = JSON.parse(File.read(file_name, encoding: 'UTF-8')).to_set
     end
 
     def load_skipped_examples_cache(cache_dir)
@@ -141,7 +144,7 @@ module RSpecTracer
 
       return unless File.file?(file_name)
 
-      @skipped_examples = JSON.parse(File.read(file_name)).to_set
+      @skipped_examples = JSON.parse(File.read(file_name, encoding: 'UTF-8')).to_set
     end
 
     def load_all_files_cache(cache_dir)
@@ -149,7 +152,7 @@ module RSpecTracer
 
       return unless File.file?(file_name)
 
-      @all_files = JSON.parse(File.read(file_name)).transform_values do |files|
+      @all_files = JSON.parse(File.read(file_name, encoding: 'UTF-8')).transform_values do |files|
         files.transform_keys(&:to_sym)
       end
     end
@@ -159,7 +162,7 @@ module RSpecTracer
 
       return unless File.file?(file_name)
 
-      @dependency = JSON.parse(File.read(file_name)).transform_values(&:to_set)
+      @dependency = JSON.parse(File.read(file_name, encoding: 'UTF-8')).transform_values(&:to_set)
     end
 
     def load_examples_coverage_cache(cache_dir)
@@ -167,7 +170,7 @@ module RSpecTracer
 
       return unless File.file?(file_name)
 
-      @examples_coverage = JSON.parse(File.read(file_name))
+      @examples_coverage = JSON.parse(File.read(file_name, encoding: 'UTF-8'))
     end
   end
 end