+++ lib/rspec_tracer/coverage_reporter.rb
@@ -72,7 +72,8 @@ module RSpecTracer
                             end
 
         line_coverage.each_pair do |line_number, strength|
-          line_coverage_dup[line_number.to_i] += strength
+          index = line_number.to_i
+          line_coverage_dup[index] = (line_coverage_dup[index] || 0) + strength
         end
 
         @coverage[file_path] = line_coverage_dup.freeze
@@ -164,7 +165,7 @@ module RSpecTracer
 
     def jruby_line_stub(file_path)
       lines = File.foreach(file_path).map { nil }
-      root_node = ::JRuby.parse(File.read(file_path))
+      root_node = ::JRuby.parse(File.read(file_path, encoding: 'UTF-8'))
 
       visitor = org.jruby.ast.visitor.NodeVisitor.impl do |_name, node|
         if node.newline?